+++ khal/settings/settings.py
@@ -39,23 +39,38 @@ SPECPATH = os.path.join(os.path.dirname(__file__), 'khal.spec')
 def find_configuration_file():
     """Return the configuration filename.
 
-    This function builds the list of paths known by khal and
-    then return the first one which exists. The first paths
-    searched are the ones described in the XDG Base Directory
-    Standard. Each one of this path ends with
+    This function builds the list of paths known by khal and then return the
+    first one which exists. The first paths searched are the ones described in
+    the XDG Base Directory Standard, e.g. ~/.config/khal/config, additionally
+    ~/.config/khal/khal.conf is searched (deprecated). All other paths end with
     DEFAULT_PATH/DEFAULT_FILE.
 
     On failure, the path DEFAULT_PATH/DEFAULT_FILE, prefixed with
     a dot, is searched in the home user directory. Ultimately,
     DEFAULT_FILE is searched in the current directory.
     """
+    # TODO re-simplify after next v0.9.0 release
     DEFAULT_FILE = __productname__ + '.conf'
     DEFAULT_PATH = __productname__
     resource = os.path.join(DEFAULT_PATH, DEFAULT_FILE)
 
     paths = []
-    paths.extend([os.path.join(path, resource)
-                  for path in xdg.BaseDirectory.xdg_config_dirs])
+    paths = [os.path.join(path, os.path.join(DEFAULT_PATH, 'config'))
+             for path in xdg.BaseDirectory.xdg_config_dirs]
+    for path in paths:
+        if os.path.exists(path):
+            return path
+
+    paths = [os.path.join(path, resource) for path in xdg.BaseDirectory.xdg_config_dirs]
+    for path in paths:
+        if os.path.exists(path):
+            logger.warning(
+                'Deprecation Warning: configuration file path `{}` will not be '
+                'supported from the next release onwards, please migrate to '
+                '`{}` or check  the documentation.'
+                ''.format(path, path.replace('khal.conf', 'config')))
+            return path
+    paths = []
     paths.append(os.path.expanduser(os.path.join('~', '.' + resource)))
     paths.append(os.path.expanduser(DEFAULT_FILE))